# contracts/openapi/platform.v1.yaml
openapi: 3.1.0
info:
  title: Academic Agent Platform API
  version: 1.0.0
  description: Web SaaS gateway API (platform-java). Frontend consumes only this contract.
servers:
  - url: /api/v1
    description: Same-origin via Nginx
...
paths:
...
  /threads:
    get:
      tags: [threads]
      summary: List threads
      operationId: listThreads
      parameters:
        - name: project_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreadListResponse"
    post:
      tags: [threads]
      summary: Create thread
      operationId: createThread
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateThreadRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thread"
  /threads/{threadId}:
    get:
      tags: [threads]
      summary: Get thread
      operationId: getThread
      parameters:
        - $ref: "#/components/parameters/ThreadId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thread"
        "404":
          $ref: "#/components/responses/ErrorResponse"
    patch:
      tags: [threads]
      summary: Rename thread
      operationId: renameThread
      parameters:
        - $ref: "#/components/parameters/ThreadId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameThreadRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thread"
  /threads/{threadId}/messages:
    get:
      tags: [threads]
      summary: List messages
      operationId: listMessages
      parameters:
        - $ref: "#/components/parameters/ThreadId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageListResponse"
    post:
      tags: [threads]
      summary: Add user message and start run
      operationId: sendMessage
      parameters:
        - $ref: "#/components/parameters/ThreadId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Run created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartRunResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
    ThreadId:
      name: threadId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  schemas:
...
    ErrorBody:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        trace_id:
          type: string
...
    Thread:
      type: object
      required: [thread_id, project_id, created_at, current_mode, lifecycle_state]
      properties:
        thread_id:
          type: string
        project_id:
          type: string
        name:
          type: string
        created_at:
          type: string
        current_mode:
          type: string
          enum: [idea_plan, experiment_design]
        lifecycle_state:
          type: string
        idea_version:
          type: integer
        impact_level:
          type: string
    ThreadSession:
      type: object
      required: [thread_id, project_id, title, created_at, status]
      properties:
        thread_id:
          type: string
        project_id:
          type: string
        title:
          type: string
        created_at:
          type: string
        status:
          type: string
        last_message_preview:
          type: string
    ThreadListResponse:
      type: object
      required: [threads]
      properties:
        threads:
          type: array
          items:
            $ref: "#/components/schemas/ThreadSession"
    CreateThreadRequest:
      type: object
      required: [project_id]
      properties:
        project_id:
          type: string
        name:
          type: string
        mode:
          type: string
          enum: [idea_plan, experiment_design]
    RenameThreadRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
    Message:
      type: object
      required: [message_id, thread_id, role, content, created_at]
      properties:
        message_id:
          type: string
        thread_id:
          type: string
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        created_at:
          type: string
    MessageListResponse:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
    SendMessageRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
    Run:
      type: object
      required: [run_id, thread_id, status, created_at]
      properties:
        run_id:
          type: string
        thread_id:
          type: string
        status:
          type: string
          enum: [queued, running, completed, failed, cancelled]
        idea:
          type: string
        artifact_id:
          type: string
        error:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
    StartRunResponse:
      type: object
      required: [run, thread]
      properties:
        run:
          $ref: "#/components/schemas/Run"
        thread:
          $ref: "#/components/schemas/Thread"
...